Extract recorder JavaScript module
